refactor: Cleanup: migrate inline #[cfg(test)] modules to sibling tests… [ORB-00402]
Cleanup: migrate inline #[cfg(test)] modules to sibling tests/ layout (incremental)
